# Short circuit Hash#find_entry for empty Hashes