@monouser7dig 2017-10-06
like what percentage of empty hashes do you expect to be present?
(seems not all that frequent to me tbh)

if it's just 1/1000 hashes that is empty there is not much to gain and if there are lots of empty hashes then there probably is something wrong in the code?