@Fryguy 2017-10-06
Updated the OP with kemal's performance numbers:

> In this case, they showed a 3-5% performance improvement: https://gitter.im/sdogruyol/kemal?at=59d66e18f7299e8f53b14208