@yxhuvud 2017-10-14
These benchmarks are quite suspect, as they basically only verify that the CPU branch prediction work.